Flyttet flere sourcefiler fra D3Plotter til D3DGraphics Flyttet ExpressionWrapper-filerne til Math/Expression MathLib.h er barberet ned til FunctionTemplate, verifyEqualtsInt og VerifyAlmostEquals VectorFunction, ComplexFunction, og alle andre functions der bruger FunctionTemplate er lagt ud i de enkelte include-filer der definerer domain- og eller range-klassen

+++ MakeGif/MakeGifDoc.cpp
@@ -95,7 +95,7 @@ void CMakeGifDoc::Dump(CDumpContext& dc) const {
 
 
 static int readGifStreamFunction(GifFileType *gf, GifByteType *buffer, int n) {
-  return ((ByteInputStream*)gf->UserData)->getBytes((BYTE*)buffer, n);
+  return (int)(((ByteInputStream*)gf->UserData)->getBytes((BYTE*)buffer, n));
 }
 
 static int writeGifStreamFunction(GifFileType *gf, const GifByteType *buffer, int n) {
